Refactored out the picture change component to new file
