--- pages/[uid]/post/[id].js
@@ -1,5 +1,0 @@
-export default function Post() {
-    return (
-        <> soon </>
-    )
-}
